refactor: extract headless Classifier into feature-classifier-core
Move the Classifier into a new `feature-classifier-core` distribution so
it can be used headlessly, without napari, Qt or magicgui. Also add a
portable, versioned model bundle format for serialization across tools.

- New `feature-classifier-core` package: Classifier (training, prediction)
  plus `export_bundle()` / `load()` using a neutral, versioned joblib
  bundle {format_version, estimator, feature_names, class_names}. `save()`
  still writes the full pickle for GUI resume; `load()` accepts both.
- Keep `napari_feature_classifier.classifier` as a re-export shim so old
  imports and existing `.clf` pickles keep resolving.
- Make the plugin's dependencies dynamic and pin feature-classifier-core to
  the exact same version via a build-time metadata hook, so the two
  lockstep packages always match (incl. hatch-vcs dev builds).
- Wire both editable packages into pixi.
